LayoutTests/fast/js/has-own-property-expected.txt

This test verifies the behaviour of Object.prototype.hasOwnProperty, as documented in ECMA-262 rev3 section 15.2.4.5.

If the test passes, you should see two 'true' outputs followed by three 'false' outputs.

anObject has own property 'foo': true
aString has own property 'length': true
anObject has own property 'bar': false
anObject has own property 'toString': false
aString has own property 'toString': false
